Split process finding logic into different files
